Create README.md
This is a work in progress. Through talks with various medical professionals, I found that while there are many software applications in their field, there are very few mobile applications that help them with their jobs. This app plans to bring all the lists they deal with on a daily basis into on fluid application at their finger tips.